Fix lib dependencies.
