+++ pywren/__init__.py
@@ -1,9 +1,12 @@
 from __future__ import absolute_import
-from pywren.wren import default_executor, wait, lambda_executor, dummy_executor, remote_executor, get_all_results
-from pywren import wrenlogging
-from pywren.version import __version__
+
 import logging
 import os
+
+from pywren.wren import *
+from pywren import wrenlogging
+from pywren.version import __version__
+
 if "PYWREN_LOGLEVEL" in os.environ:
     log_level = os.environ['PYWREN_LOGLEVEL']
     wrenlogging.default_config(log_level)
